Create Technical documentation page
 technical documentation page can be made if you are comfortable with the foundations of HTML, CSS, and JavaScript. The primary objective of this project is to create a page of technical documentation where users can click on any topic on the left side of the page and have the appropriate content load on the right. 

There is nothing extravagant about the project; it is just a simple, straightforward page of technical documentation. To build this HTML project, you must divide the webpage into two parts. On the right side will be the documentation (description) for each topic, and on the left side will be a menu listing each topic in alphabetical order. The click function can be added using Javascript or CSS bookmarks.