Create Landing page
Another excellent HTML project example for beginners is this one, which you can create and include in your portfolio. A landing page is a single web page that promotes one or more goods and services while directing site visitors to carry out various actions, such as making a purchase, signing up for a newsletter, etc. You need to combine CSS with HTML to develop even the most basic landing page. 

A landing page must be appealing and purposeful if it hopes to persuade visitors to take some sort of action. Your styling of the landing page should centre on selecting the right colour schemes, page layout, etc. Additionally, you must ensure that the page's elements do not overlap or be put incorrectly.