Create Parallax website
A newcomer who is familiar with the fundamentals of HTML can create a parallax website in a single day! A parallax website enables you to scroll up and down the page to view the various elements of an image that is fixed in the backdrop. It creates a stunning and distinctive impact on a website. 

Section the page into three to four parts before beginning to develop the parallax site. Select a few background pictures, place them in the proper portions of the page with the pertinent text, adjust the margin and padding, and add a background position. CSS can be used to add additional fashionable components to the page. It can be considered one of the best HTML CSS projects for beginners.