Create A Simple Webpage with Animation
One of the finest methods to increase the interaction and appeal of a web page is through animation. A web page with certain animations can be made as an HTML project. GIFs and photos can be used to provide animation effects to a website. 

Typically, CSS and JavaScript are used to create the basic animations that we see on many websites. However, you may use HTML to create a web page and add a few basic animations to it in order to learn more about the markup language's potential.