Create Personal Portfolio page
You must have a strong understanding of HTML5 and CSS3 in order to build a personal portfolio page. The basic information for a work portfolio, such as your name and image, projects, specialised talents, and interests, will be included on the web page you make for this project. If you'd like, you may also host the entire portfolio on GitHub using your GitHub account and include your resume. 

A header and footer section should be present on the portfolio page. A menu highlighting your contact details, employment history, and personal information will be included in the header area. You can provide a brief description of your professional background and hobbies along with your photo at the top of the page. You can add a few work samples after this summary.