Create Music store page
For music enthusiasts, a music store page is the ideal testing ground. You must be an expert in HTML5 and CSS3 in order to construct this page. 

The first step in creating a music page is to include an appropriate background image and a brief description of the content. Different menus that list songs according to attributes like genre, year, singer, album, etc. can be found in the header part of the page. 

You will need to include buttons for start, stop, rewind/forward, and other functions. For the available song collection, include pertinent links and pictures. Contact information, links for registration, in-store  

purchases, membership plans, and trial alternatives may all be found in the bottom.