Create Tribute Page
This is another of those HTML practice projects that are ideal for beginners. With this tribute page, you’ll combine your HTML skills and your passion for a person or topic.

This tribute page project requires you to edit the page title and description, add sections for notable quotes and images, and also create a biography section. 

This is another project that’s good for using HTML elements along with CSS to enhance styling and appearance. You’ll be able to practice using paragraph elements and key CSS styles for box-sizing, margins, and padding with your tribute page.