Create Conference Webpage
This is one of the easier beginner HTML projects that are great for experimenting with your HTML skills, as it involves building a static page to display detailed information about an upcoming event. You’ll also be able to combine HTML and CSS, which is good for honing your UX skills.

Feel free to experiment with this project by dividing the page into smaller sections with div elements. This is also helpful in making the page more organized. You’ll also be able to create header and footer semantic tags, along with a menu display. 

When it comes to CSS, play around with color choices for different sections, and also try out different font types and colors to get the perfect theme for your website.