Create Restaurant Website
This HTML project involves creating a stunning website for a restaurant to demonstrate your command of HTML and CSS, and as a more advanced HTML project, you can expect the web page layout to be more challenging than the previous projects. 

Some of the key elements of this project involve more advanced CSS skills, including a CSS layout grid to position foods and beverages on the page. You’ll also need to show prices and images, not to mention style choices regarding the right blend of colors and font styles.

If you’re interested in front-end web design, this project boosts your skills. You can even include a picture gallery with sliding images to enhance this website’s appearance. If you want to take this a step further, research ways to make your website responsive and mobile-friendly.