Create Music Player
f you're a music lover, this is one of the most fun, relatively advanced HTML projects. You’ll use a blend of HTML, CSS, and JavaScript skills to make the most of this project.

Some of the key aspects of this project include adding important music control buttons like play, stop, rewind, etc. You’ll also need to add a stylish backdrop or thematic color choice to ensure it looks attractive.

In terms of HTML, you’ll need to create class containers and divs and then blend this with CSS for styling and jQuery for music player functionalities.